scraper-service/pyproject.toml
[project]
name = "scraper-service"
version = "0.1.0"
description = "Standalone scraping microservice with unified fetch/parse engine"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.41.0",
    "httpx>=0.28.1",
    "curl-cffi>=0.14.0",
    "playwright>=1.58.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "selectolax>=0.4.6",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
    "PyMuPDF>=1.27.1",
    "pytesseract>=0.3.13",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "cachetools>=7.0.1",
    "tldextract>=5.3.1",
    "matrx-utils>=1.0.4",
    "markdownify>=0.14.1",
    "tabulate>=0.9.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.11",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-httpx>=0.35.0",
    "rich>=14.3.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
